+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".LauncherActivity"
+            android:name=".launcher.LauncherActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
