+++ app/src/main/java/com/example/memowithtags/common/network/AuthInterceptor.kt
@@ -1,4 +1,4 @@
-package com.example.memowithtags.Network
+package com.example.memowithtags.common.network
 
 import okhttp3.Interceptor
 import okhttp3.Response
