+++ app/src/main/java/com/example/memowithtags/launcher/LauncherActivity.kt
@@ -1,10 +1,10 @@
-package com.example.memowithtags
+package com.example.memowithtags.launcher
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import com.example.memowithtags.Viewmodels.LauncherViewModel
+import com.example.memowithtags.launcher.viewModel.LauncherViewModel
 import com.example.memowithtags.login.LoginActivity
 import com.example.memowithtags.mainMemo.MainActivity
 import dagger.hilt.android.AndroidEntryPoint
