+++ app/src/main/java/com/example/memowithtags/launcher/viewModel/LauncherViewModel.kt
@@ -1,6 +1,7 @@
-package com.example.memowithtags.Viewmodels
+package com.example.memowithtags.launcher.viewModel
 
 import androidx.lifecycle.ViewModel
+import com.example.memowithtags.signup.repository.AuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
