+++ app/src/main/java/com/example/memowithtags/signup/fragments/SignupStep1Fragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.memowithtags.Network.SendEmailRequest
+import com.example.memowithtags.common.network.SendEmailRequest
 import com.example.memowithtags.R
 import com.example.memowithtags.databinding.FragmentSignupStep1Binding
 import com.example.memowithtags.network.ApiClient
