+++ app/src/main/java/com/example/memowithtags/signup/fragments/SignupStep2Fragment.kt
@@ -9,7 +9,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.memowithtags.Network.VerifyEmailRequest
+import com.example.memowithtags.common.network.VerifyEmailRequest
 import com.example.memowithtags.R
 import com.example.memowithtags.databinding.FragmentSignupStep2Binding
 import com.example.memowithtags.network.ApiClient
