+++ app/src/main/java/com/example/memowithtags/signup/viewModel/SignupViewModel.kt
@@ -3,9 +3,9 @@ package com.example.memowithtags.signup.viewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.memowithtags.Network.SignupRequest
-import com.example.memowithtags.Network.SignupResponse
-import com.example.memowithtags.Viewmodels.AuthRepository
+import com.example.memowithtags.common.network.SignupRequest
+import com.example.memowithtags.common.network.SignupResponse
+import com.example.memowithtags.signup.repository.AuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import retrofit2.Call
 import retrofit2.Callback
